pyproject.toml
[project]
name = "nasa-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.12.15",
    "fastmcp>=2.12.2",
    "httpx>=0.28.1",
    "mcp[cli]>=1.13.1",
    "pandas>=2.3.2",
    "pillow>=11.3.0",
    "python-dotenv>=1.1.1",
    "pytz>=2025.2",
    "requests>=2.32.5",
    "ruff>=0.12.12",
]
